refactor(restify): extract AuthorizationService
